BayesClassifier generics

Generic functions implemented for the BayesClassifier class

Usage

# S3 method for BayesClassifier
print(x, ...)

# S3 method for BayesClassifier
summary(object, ...)

# S3 method for BayesClassifier
coef(object, ...)

# S3 method for BayesClassifier
dimnames(x)

# S3 method for BayesClassifier
dim(x)

# S3 method for BayesClassifier
levels(x)

# S3 method for BayesClassifier
length(x)

# S3 method for BayesClassifier
nobs(object, ...)

# S3 method for BayesClassifier
formula(x, ...)

# S3 method for BayesClassifier
logLik(object, ...)

# S3 method for BayesClassifier
AIC(object, ..., k = 2)

# S3 method for BayesClassifier
BIC(object, ...)

Arguments

x

a BayesClassifier object

...

currently unused

object

a BayesClassifier object

k

penalty term; k=2 in classical AIC

Value

The form of the returned values of the generic functions for class ´BayesClassifier´ are as follows:

´print´ and ´summary´ return a console output only

´coef´ returns the list of parameters, see BayesClassifier

´dimnames´ and ´levels´ return a vector containing the feature names and class names, respectively

´dim´, ´length´ and ´nobs´ return a scalar representing the number of features, the number of classes and the number of training observations, respectively

´formula´ returns the formula object used to call the BayesClassifier

´logLik´, ´AIC´ and ´BIC´ return the log-likelihood, the AIC and the BIC values, respectively

Functions

print(BayesClassifier): prints the BayesClassifier

summary(BayesClassifier): summarizes the BayesClassifier

coef(BayesClassifier): returns the list of model parameters in the BayesClassifier

dimnames(BayesClassifier): returns the names the features used in the BayesClassifier

dim(BayesClassifier): returns the dimensionality of the feature space used in the BayesClassifier

levels(BayesClassifier): returns the names of the classes in the BayesClassifier

length(BayesClassifier): returns the number of classes in the BayesClassifier

nobs(BayesClassifier): returns the number of training observations used for the BayesClassifier

formula(BayesClassifier): returns the formula of the BayesClassifier

logLik(BayesClassifier): returns the log-likelihood of the BayesClassifier

AIC(BayesClassifier): returns the Akaike Information Criterion for the BayesClassifier

BIC(BayesClassifier): returns the Bayesian Information Criterion for the BayesClassifier